Rename ThresholdsParameter to ThresholdsField
